[feat] Created Survey Results page
Created page, and added graph for baseline

Fix build failure

All graphs added
