+++ src/components/Review/WeekInReviewReview.tsx
@@ -1,10 +1,10 @@
 import { IonIcon, IonSpinner } from "@ionic/react";
-import { get, limitToLast, orderByKey, query, ref, serverTimestamp, set } from "firebase/database";
+import { ref, serverTimestamp, set } from "firebase/database";
 import { useEffect, useState } from "react";
 import { auth, db } from "../../firebase";
-import { AnyMap, checkKeys, decrypt, getDateFromLog, toast } from "../../helpers";
+import { AnyMap, BaselineStates, BASELINE_GRAPH_CONFIG, calculateBaseline, parseSurveyHistory, toast } from "../../helpers";
 import history from "../../history";
-import Screener, { GraphConfig, Priority } from "../../screeners/screener";
+import Screener, { Priority } from "../../screeners/screener";
 import SwiperType, { Pagination } from "swiper";
 import { Swiper, SwiperSlide } from "swiper/react"
 import "swiper/css";
@@ -12,18 +12,11 @@ import "swiper/css/pagination";
 import { chevronBackOutline, chevronForwardOutline } from "ionicons/icons";
 import { useAuthState } from "react-firebase-hooks/auth";
 import SurveyGraph from "./SurveyGraph";
-import ldb from "../../db";
-import { DateTime } from "luxon";
 import EndSpacer from "../EndSpacer";
 
 interface Props {
     primary: Screener,
     secondary: Screener
-}
-
-enum BaselineStates {
-    NOT_STARTED,
-    NOT_ENOUGH_DATA
 }
 
 const WeekInReviewReview = ({ primary, secondary }: Props) => {
@@ -32,14 +25,6 @@ const WeekInReviewReview = ({ primary, secondary }: Props) => {
     const [user] = useAuthState(auth);
     const [surveyHistory, setSurveyHistory] = useState<AnyMap | undefined>(undefined);
     const [baselineGraph, setBaselineGraph] = useState<AnyMap[] | BaselineStates>(BaselineStates.NOT_STARTED);
-    const baselineGraphConfig: GraphConfig = {
-        yAxisLabel: "baseline (-5 to 5 scale)",
-        lines: [{
-            key: "Mood",
-            color: "#955196"
-        }]
-    };
-    const BASELINE_DAYS = 14;
 
     const finish = async () => {
         if (loading) return;
@@ -60,80 +45,11 @@ const WeekInReviewReview = ({ primary, secondary }: Props) => {
         });
 
     useEffect(() => {
-        if (!user) return;
-        const keys = checkKeys();
-        get(query(ref(db, `${user.uid}/surveys`), orderByKey(), limitToLast(100))).then(snap => {
-            let val = snap.val();
-            for (let key in val) {  
-                if (typeof val[key]["results"] === "string") {
-                    val[key]["results"] = JSON.parse(decrypt(val[key]["results"], `${keys.visibleKey}${keys.encryptedKeyVisible}`));
-                }
-            }
-
-            setSurveyHistory(val);
-        });
+        parseSurveyHistory(user, setSurveyHistory);
     }, [user]);
 
     useEffect(() => {
-        (async () => {
-            const logs = await ldb.logs.where("timestamp").above(DateTime.now().minus({ years: 1 }).toMillis()).toArray();
-            let currentDate = getDateFromLog(logs[0]);
-            let ptr = 0;
-            let perDayDates = [];
-            let perDayAverages = [];
-            const now = DateTime.local();
-            // Aggregate average mood data per day
-            while (currentDate < now) {
-                let todaySum = 0;
-                let ctr = 0;
-                while (
-                    ptr < logs.length 
-                    && logs[ptr].day === currentDate.day 
-                    && logs[ptr].month === currentDate.month 
-                    && logs[ptr].year === currentDate.year
-                ) {
-                    if (logs[ptr].average === "average") {
-                        todaySum += logs[ptr].mood;
-                        ++ctr;
-                    }
-                    ++ptr;
-                }
-                perDayAverages.push(ctr === 0 ? 0 : todaySum / ctr);
-                perDayDates.push(currentDate.toFormat("LLL d"))
-                currentDate = currentDate.plus({"days": 1});
-            }
-
-            if (perDayAverages.length <= BASELINE_DAYS) {
-                setBaselineGraph(BaselineStates.NOT_ENOUGH_DATA);
-                return;
-            }
-
-            let sum = 0;
-            let i = 0;
-            for (i = 0; i < BASELINE_DAYS; ++i) {
-                sum += perDayAverages[i];
-            }
-
-            // Start average with base 14 days
-            let baseline = [];
-            baseline.push({
-                date: perDayDates[i - 1],
-                Mood: sum / BASELINE_DAYS
-            });
-
-            // Calculate rolling average to end of list
-            while (i < perDayAverages.length) {
-                sum -= perDayAverages[i - BASELINE_DAYS];
-                sum += perDayAverages[i];
-                baseline.push({
-                    date: perDayDates[i],
-                    Mood: sum / BASELINE_DAYS
-                });
-                ++i;
-            }
-
-            setBaselineGraph(baseline);
-        })();
+        calculateBaseline(setBaselineGraph);
     }, []);
     
     return <div className="center-summary container">
@@ -171,7 +87,7 @@ const WeekInReviewReview = ({ primary, secondary }: Props) => {
                     <div className="title">Your baseline</div>
                     <div className="text-center screener-slide">
                         { typeof baselineGraph === "object" && <>
-                            <SurveyGraph data={baselineGraph} graphConfig={baselineGraphConfig} />
+                            <SurveyGraph data={baselineGraph} graphConfig={BASELINE_GRAPH_CONFIG} />
                             <p>
                                 Remember how every time you mood log, we ask you whether you're feeling below, at, 
                                 or above average? Well, here's what that's used for &mdash; your baseline. Your baseline
