+++ src/pages/Summary.css
@@ -135,3 +135,7 @@
 .menu > .move-rest-down {
     margin-top: auto;
 }
+
+.head2 {
+    font-size: 20px;
+}
